refactor(admin): move stats dashboard to its own /admin/stats tab
Splits the dashboard out of the Users page into a dedicated route alongside
Users and Reports, and points the /admin index redirect at it.
