+++ apps/web/src/routes/admin.index.tsx
@@ -2,6 +2,6 @@ import { createFileRoute, redirect } from "@tanstack/react-router"
 
 export const Route = createFileRoute("/admin/")({
   beforeLoad: () => {
-    throw redirect({ to: "/admin/users" })
+    throw redirect({ to: "/admin/stats" })
   },
 })
